feat: 판매자 물품 조회 기능 구현 완

+++ src/main/java/com/homesweet/homesweetback/domain/product/product/controller/response/SkuStockResponse.java
@@ -10,7 +10,7 @@ import java.util.List;
  */
 public record SkuStockResponse(
         Long skuId,
-        Integer stockQuantity,
+        Long stockQuantity,
         Integer priceAdjustment,
         List<OptionCombinationResponse> options
 ) {
